split mesh_util into numba
